+++ src/storage/dictionary/dictionary_encoder.h
@@ -1,4 +1,3 @@
-
 #pragma once
 
 #include "dictionary.h"
